Move `core::config` macros into a separate file
